Refactor: Extract MessageWidget to its own file
