+++ lib/cases/investigation_widget.dart
@@ -1,4 +1,3 @@
-import 'dart:convert';
 import 'package:common/common.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -6,6 +5,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:providers/firestore.dart';
 import 'package:widgets/col_stream_widget.dart';
 import 'package:widgets/doc_stream_widget.dart';
+import 'message_widget.dart'; 
 
 class InvestigationWidget extends ConsumerWidget {
   final DR caseDocRef;
@@ -117,38 +117,3 @@ class SendMessageWidget extends ConsumerWidget {
   }
 }
 
-class MessageWidget extends ConsumerWidget {
-  final DR messageDocRef;
-
-  MessageWidget(this.messageDocRef);
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return DocStreamWidget(
-      key: ValueKey(messageDocRef.id),
-      docSP(messageDocRef.path),
-      (context, messageDoc) => Card(
-        child: ListTile(
-          trailing: IconButton(
-            onPressed: () => messageDocRef.delete(),
-            icon: Icon(Icons.delete),
-          ),
-          leading: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Text(messageDoc.data()?['role'] ?? ''),
-              
-            ],
-          ),
-          title: Container(
-            width: 200,
-            child: ConstrainedBox(
-              constraints: BoxConstraints(minHeight: 50),
-              child: Text(messageDoc.data()?['content'] ?? ''),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
